requirements.txt
alabaster==0.7.13
anndata==0.10.5.post1
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.4.1
arrow==1.3.0
astroid==3.0.1
asttokens==2.4.1
async-lru==2.0.4
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
biopython==1.83
black==23.11.0
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cookiecutter==2.4.0
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
docopt==0.6.2
docstring-to-markdown==0.13
docutils==0.20.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.18.1
flake8==6.1.0
fonttools==4.44.1
fqdn==1.5.1
goatools==1.3.11
h5py==3.10.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.19.1
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
keyring==24.3.0
kiwisolver==1.4.5
kneed==0.8.5
kneefinder==0.0.2
legacy-api-wrap==1.4
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.1.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.0
numpy==1.26.2
numpydoc==1.6.0
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pydot==1.4.2
pyflakes==3.1.0
Pygments==2.16.1
pylint==3.0.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pynndescent==0.5.11
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyreader==0.0.1
pyreadr==0.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
python-slugify==8.0.1
pytoolconfig==1.2.6
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==3.2.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rope==1.11.0
rpds-py==0.12.0
Rtree==1.1.0
scanpy==1.9.5
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.11.2
Send2Trash==1.8.2
session_info==1.0.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
spyder==5.5.0
spyder-kernels==2.5.0
stack-data==0.6.3
statannot==0.2.3
statannotations==0.6.0
statsmodels==0.14.0
stdlib-list==0.10.0
tabulate==0.9.0
terminado==0.18.0
text-unidecode==1.3
textdistance==4.6.0
threadpoolctl==3.2.0
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
ujson==5.8.0
umap-learn==0.5.5
uri-template==1.3.0
urllib3==2.1.0
venn==0.1.3
watchdog==3.0.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whatthepatch==1.0.5
widgetsnbextension==4.0.9
XlsxWriter==3.1.9
yapf==0.40.2
zipp==3.17.0
